dmarc summaries connections refactored
