+++ api-js/src/dmarc-summaries/objects/__tests__/detail-tables.test.js
@@ -9,10 +9,10 @@ import {
   spfFailureLoaderConnectionsBySumId,
 } from '../../loaders'
 import { detailTablesType } from '../detail-tables'
-import { dkimFailureConnection } from '../dkim-failure-table'
-import { dmarcFailureConnection } from '../dmarc-failure-table'
-import { fullPassConnection } from '../full-pass-table'
-import { spfFailureConnection } from '../spf-failure-table'
+import { dkimFailureConnection } from '../dkim-failure-table-connection'
+import { dmarcFailureConnection } from '../dmarc-failure-table-connection'
+import { fullPassConnection } from '../full-pass-table-connection'
+import { spfFailureConnection } from '../spf-failure-table-connection'
 
 const { DB_PASS: rootPass, DB_URL: url } = process.env
