+++ api-js/src/dmarc-summaries/objects/dkim-failure-table-connection.js
@@ -1,0 +1,16 @@
+import { GraphQLInt } from 'graphql'
+import { connectionDefinitions } from 'graphql-relay'
+
+import { dkimFailureTableType } from './dkim-failure-table'
+
+export const dkimFailureConnection = connectionDefinitions({
+  name: 'DkimFailureTable',
+  nodeType: dkimFailureTableType,
+  connectionFields: () => ({
+    totalCount: {
+      type: GraphQLInt,
+      description: 'The total amount of dkim failure the user has access to.',
+      resolve: ({ totalCount }) => totalCount,
+    },
+  }),
+})
