+++ api-js/src/dmarc-summaries/objects/index.js
@@ -2,7 +2,12 @@ export * from './category-percentages'
 export * from './category-totals'
 export * from './detail-tables'
 export * from './dkim-failure-table'
+export * from './dkim-failure-table-connection'
 export * from './dmarc-failure-table'
+export * from './dmarc-failure-table-connection'
 export * from './dmarc-summary'
+export * from './dmarc-summary-connection'
 export * from './full-pass-table'
+export * from './full-pass-table-connection'
 export * from './spf-failure-table'
+export * from './spf-failure-table-connection'
